// types.ts
export type PostType = {
  allow_reply: boolean;
  body: string;
  createdAt: string;
  description: string;
  id: number;
  is_blocked: boolean;
  is_reported: number;
  likes: number;
  public: boolean;
  scraps: number;
  slug: string;
  thumbnail: string;
  title: string;
  updatedAt: string;
  users_id: number;
  weather: string;
  user: string;
};